backend/requirements.txt
# Mamaope Legal AI - Enhanced Security Requirements
# Medical software with HIPAA/GDPR/ISO 13485 compliance

# Core FastAPI and web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
starlette==0.41.3
pydantic==2.10.4
pydantic-settings==2.6.1

# Database and ORM
sqlalchemy==2.0.36
alembic==1.16.5
psycopg2-binary==2.9.10

# Authentication and Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.20
cryptography==46.0.1
PyJWT==2.9.0
email-validator==2.2.0

# AI and ML
google-generativeai==0.8.3
google-genai==0.5.0
vertexai==1.71.1
openai==1.70.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28

# Vector Database
pymilvus==2.4.9
milvus-lite==2.4.12

# Data Processing
pandas==2.3.2
numpy==1.26.4
python-docx==1.1.2
PyPDF2==3.0.1
PyMuPDF==1.26.4

# Cloud Services
boto3==1.35.90
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.71.1

# Security and Validation
bleach==6.2.0
defusedxml==0.7.1
python-dotenv==1.0.1

# Monitoring and Logging
structlog==24.4.0
coloredlogs==15.0.1

# Testing and Quality
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.28.0

# Code Quality and Security
bandit==1.7.10
safety==3.2.1
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Development and Debugging
tenacity==9.0.0
rich==13.9.4
click==8.1.7

# HTTP and Networking
httpx==0.28.0
aiohttp==3.11.9
requests==2.32.3

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
typing-extensions==4.12.2